# packages/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiSessionContext": {
        "properties": {
          "ended_at": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "session_id": {
            "type": "string"
          },
          "started_at": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "workspace_id": {
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "session_id",
          "started_at"
        ],
        "type": "object"
      },
      "ApplicationInfo": {
        "properties": {
          "bundleId": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "path"
        ],
        "type": "object"
      },
      "ApplicationsResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/ApplicationInfo"
            },
            "type": "array"
          },
          "count": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "applications",
          "count"
        ],
        "type": "object"
      },
...
      "EnqueueMessageResponse": {
        "properties": {
          "accepted_at": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "context": {
            "$ref": "#/components/schemas/ApiSessionContext"
          },
          "run_id": {
            "type": "string"
          }
        },
        "required": [
          "context",
          "run_id",
          "accepted_at"
        ],
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "info": {
            "$ref": "#/components/schemas/MessageInfo"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/MessagePart"
            },
            "type": "array"
          }
        },
        "required": [
          "info",
          "parts"
        ],
        "type": "object"
      },
      "MessageInfo": {
        "properties": {
          "completedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "role",
          "createdAt"
        ],
        "type": "object"
      },
      "MessagePart": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/TextPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "text"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReasoningPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "reasoning"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "tool"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExtensionPart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "extension"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FilePart"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "file"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
...
      "OpenApplicationRequest": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "OpenApplicationResponse": {
        "properties": {
          "id": {
            "type": "string"
          },
          "opened": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "opened",
          "id"
        ],
        "type": "object"
      },
      "OpenExtensionRequest": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "PartBase": {
        "properties": {
          "id": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "messageId"
        ],
        "type": "object"
      },
      "PollResponse": {
        "properties": {
          "authorization_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "ReasoningPart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "text": {
                "type": "string"
              }
            },
            "required": [
              "text"
            ],
            "type": "object"
          }
        ]
      },
      "RecordMessageRequest": {
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/SessionCommandInputPart"
            },
            "type": "array"
          }
        },
        "required": [
          "parts"
        ],
        "type": "object"
      },
...
      "SessionCommandInputPart": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionCommandTextPartInput"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "text"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionCommandExtensionPartInput"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "extension"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionCommandFilePartInput"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "file"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
      "SessionCommandTextPartInput": {
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "type": "object"
      },
...
      "StartFlowRequest": {
        "properties": {
          "state": {
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "type": "object"
      },
      "StartFlowResponse": {
        "properties": {
          "redirect_uri": {
            "type": "string"
          }
        },
        "required": [
          "redirect_uri"
        ],
        "type": "object"
      },
...
      "TextPart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PartBase"
          },
          {
            "properties": {
              "text": {
                "type": "string"
              }
            },
            "required": [
              "text"
            ],
            "type": "object"
          }
        ]
      },
...
    }
  },
  "info": {
    "description": "AI-native command bar for macOS",
    "license": {
      "name": ""
    },
    "title": "Cocommand API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/events": {
      "get": {
        "description": "Subscribe to real-time server events via SSE. Optionally filter by session_id.",
        "operationId": "stream_events",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "SSE stream of CoreEvent payloads"
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/extension/{extension_id}/invoke/{tool_id}": {
      "post": {
        "description": "Looks up the extension and tool by ID, executes the tool with the provided\nJSON input, and returns a standardised `{ ok, data }` / `{ ok, error }` envelope.",
        "operationId": "invoke_tool",
        "parameters": [
          {
            "description": "Extension identifier",
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Tool identifier",
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tool execution result envelope with { ok, data }"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "POST /extension/{extension_id}/invoke/{tool_id}",
        "tags": [
          "extensions"
        ]
      }
    },
    "/oauth/callback": {
      "get": {
        "operationId": "callback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {}
            },
            "description": "HTML page confirming authorization success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/poll": {
      "get": {
        "operationId": "poll_flow",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "wait",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/start": {
      "post": {
        "operationId": "start_flow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartFlowResponse"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/tokens/{ext}/{provider}": {
      "delete": {
        "operationId": "delete_tokens",
        "parameters": [
          {
            "description": "Extension identifier",
            "in": "path",
            "name": "ext",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "OAuth provider name",
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens deleted successfully"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "oauth"
        ]
      },
      "get": {
        "operationId": "get_tokens",
        "parameters": [
          {
            "description": "Extension identifier",
            "in": "path",
            "name": "ext",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "OAuth provider name",
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stored token JSON"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "oauth"
        ]
      },
      "post": {
        "operationId": "set_tokens",
        "parameters": [
          {
            "description": "Extension identifier",
            "in": "path",
            "name": "ext",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "OAuth provider name",
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tokens stored successfully"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "oauth"
        ]
      }
    },
    "/session/command": {
      "get": {
        "description": "Load message history for the current session.",
        "operationId": "session_command_history",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Current session message history"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/command": {
      "post": {
        "description": "Enqueue a message for asynchronous session runtime processing.",
        "operationId": "session_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueMessageResponse"
                }
              }
            },
            "description": "Message accepted and enqueued"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/context": {
      "get": {
        "operationId": "session_context",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSessionContext"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/workspace/extension/system/applications": {
      "get": {
        "operationId": "list_applications",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationsResponse"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "system"
        ]
      }
    },
    "/workspace/extension/system/applications/open": {
      "post": {
        "operationId": "open_application",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenApplicationResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "system"
        ]
      }
    },
...
    "/workspace/extensions/open": {
      "post": {
        "operationId": "open_extension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenExtensionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tool output from the extension's open tool"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "extensions"
        ]
      }
    }
  },
...
}